Source code for 87 version

+++ Debtor/SerialToOrderLinePage.xaml.cs
@@ -79,6 +79,7 @@ namespace UnicontaClient.Pages.CustomPage
             InitializeComponent();
             dcorderlineMaster = master as DCOrderLineClient;
             Comp = api.CompanyEntity;
+            api.AllowBackgroundCrud = true;
             itemCache = Comp.GetCache(typeof(InvItem));
             invItemMaster = itemCache.Get(dcorderlineMaster._Item) as InvItem;
             SetRibbonControl(localMenu, dgLinkedGrid);
@@ -281,15 +282,26 @@ namespace UnicontaClient.Pages.CustomPage
 
         private async void UnlinkRows()
         {
+            var dcorderlineMaster = this.dcorderlineMaster;
+            UnicontaBaseEntity orgMaster;
             dgLinkedGrid.SelectedItem = null;
             dgUnlinkedGrid.SelectedItem = null;
             var dcolSerieBatchList = new List<DCOrderLineSerieBatch>();
             var linkedRows = dgLinkedGrid.ItemsSource as List<SerialToOrderLineClient>;
-            if (linkedRows == null)
+            if (linkedRows == null || linkedRows.Count == 0)
             {
+                if (dcorderlineMaster._SerieBatchMarked && dcorderlineMaster.RowId != 0)
+                {
+                    orgMaster = StreamingManager.Clone((UnicontaBaseEntity)dcorderlineMaster);
+                    dcorderlineMaster.SerieBatch = null;
+                    dcorderlineMaster.SerieBatchMarked = false;
+                    api.UpdateNoResponse(orgMaster, (UnicontaBaseEntity)dcorderlineMaster);
+                }
+                dcorderlineMaster.SerieBatch = null;
                 dcorderlineMaster.SerieBatchMarked = false;
                 return;
             }
+
             bool AllRemoved = true;
             foreach (var row in linkedRows)
             {
@@ -304,12 +316,25 @@ namespace UnicontaClient.Pages.CustomPage
                 else
                     AllRemoved = false;
             }
-            var err = await api.Delete(dcolSerieBatchList);
+
+            ErrorCodes err;
+            if (dcorderlineMaster._SerieBatch != null && dcorderlineMaster.RowId != 0)
+            { 
+                orgMaster = StreamingManager.Clone((UnicontaBaseEntity)dcorderlineMaster);
+                dcorderlineMaster.SerieBatch = null;
+                if (AllRemoved)
+                    dcorderlineMaster.SerieBatchMarked = false;
+                var Updates = new Uniconta.API.System.CrudAPI.UpdatePair() { loaded = orgMaster, modified = (UnicontaBaseEntity)dcorderlineMaster };
+                err = await api.MultiCrud(null, new[] { Updates }, dcolSerieBatchList);
+            }
+            else
+                err = await api.Delete(dcolSerieBatchList);
             if (err != ErrorCodes.Succes)
                 UtilDisplay.ShowErrorCode(err);
             else
             {
                 InitQuery();
+                dcorderlineMaster.SerieBatch = null;
                 if (AllRemoved)
                     dcorderlineMaster.SerieBatchMarked = false;
             }
@@ -343,6 +368,7 @@ namespace UnicontaClient.Pages.CustomPage
                 return;
 
             UnicontaBaseEntity orgMaster = null;
+            var dcorderlineMaster = this.dcorderlineMaster;
 
             var _UseSerial = this.invItemMaster._UseSerial;
             List<DCOrderLineSerieBatch> olSerieBatchList = new List<DCOrderLineSerieBatch>();
@@ -362,14 +388,34 @@ namespace UnicontaClient.Pages.CustomPage
                 olSerieBatchList.Add(olSerieBatch);
                 if (row._Warehouse != null && (dcorderlineMaster._Warehouse != row._Warehouse || dcorderlineMaster._Location != row._Location))
                 {
-                    orgMaster = StreamingManager.Clone((UnicontaBaseEntity)dcorderlineMaster);
+                    if (orgMaster == null && dcorderlineMaster.RowId != 0)
+                        orgMaster = StreamingManager.Clone((UnicontaBaseEntity)dcorderlineMaster);
                     dcorderlineMaster.Warehouse = row._Warehouse;
                     dcorderlineMaster.Location = row._Location;
+                    if (markedList.Count == 1)
+                        dcorderlineMaster.SerieBatch = row._Number;
+                    else
+                        dcorderlineMaster.SerieBatch = null;
+                }
+                else
+                {
+                    string nr = dcorderlineMaster._SerieBatch, newNr;
+                    if (markedList.Count == 1)
+                        newNr = row._Number;
+                    else
+                        newNr = null;
+                    if (newNr != nr)
+                    {
+                        if (orgMaster == null && dcorderlineMaster.RowId != 0)
+                            orgMaster = StreamingManager.Clone((UnicontaBaseEntity)dcorderlineMaster);
+                        dcorderlineMaster.SerieBatch = newNr;
+                    }
                 }
             }
             ErrorCodes err;
             if (orgMaster != null)
             {
+                dcorderlineMaster.SerieBatchMarked = true;
                 var Updates = new Uniconta.API.System.CrudAPI.UpdatePair() { loaded = orgMaster, modified = (UnicontaBaseEntity)dcorderlineMaster };
                 err = await api.MultiCrud(olSerieBatchList, new[] { Updates }, null);
             }
@@ -420,14 +466,16 @@ namespace UnicontaClient.Pages.CustomPage
             var company = api.CompanyEntity;
             if (!company.Location || !company.Warehouse)
             {
-//                Location.Visible = Location.ShowInColumnChooser = false;
                 LocationCol.Visible = LocationCol.ShowInColumnChooser = false;
             }
+            else
+                LocationCol.ShowInColumnChooser = true;
             if (!company.Warehouse)
             {
-//                Warehouse.Visible = Warehouse.ShowInColumnChooser = false;
                 WarehouseCol.Visible = WarehouseCol.ShowInColumnChooser = false;
             }
+            else
+                WarehouseCol.ShowInColumnChooser = true;
         }
 
         public override bool CheckIfBindWithUserfield(out bool isReadOnly, out bool useBinding)
